Rewrote the SMS service abstraction so that it actually works.
It's now significantly uglier than before, but I haven't really been able to get around that.

Basically, you create a service for your SMS implementation, and you set SmsServiceFactory to return that service. But to receive SMS, you also probably need to create a servlet in the servlets package, that receives the SMS and calls some (non-service-interface) method on your service, which can then go ahead and call onSmsReceived on any registered listeners.

Look at AppengineSmsService/AppengineSmsReceiverServlet for an example.